refactor(tasks): move shared board modules out of home
Generated-By: PostHog Code
Task-Id: 1fbd8776-9d5b-4e74-9629-a6bdb8d3b78f
